refactor : writingForm to writingDTO
